refactor: FAQ 아코디언 컴포넌트 분리
- FAQ 페이지에서 아코디언 로직을 FaqAccordion 컴포넌트로 분리
- 인라인 드롭다운 SVG를 기존 dropdown asset 사용으로 변경
- 공지사항 날짜 포맷 유틸 함수 추가
