+++ apps/web/src/common/util/formatDate.ts
@@ -34,3 +34,12 @@ export function formatToMonthDay2(input: string | null): string {
 
   return `${month}.${day}`;
 }
+
+export const formatNoticeDate = (createdAt: string) =>
+  new Intl.DateTimeFormat('ko-KR', {
+    year: 'numeric',
+    month: '2-digit',
+    day: '2-digit',
+  })
+    .format(new Date(createdAt))
+    .replace(/\s/g, '');
